requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.28.0
statsmodels>=0.14.0 